platform/NVMem: fix _plat__NvEnable return codes
Most of the returns in _plat__NvEnable are inverted (except the earliest
one). This didn't result in bad behavior, as the library callers always
got the correct early return, and the TA callers assumed inverted
behavior, but it makes the code difficult to follow. Replace with macros
to avoid confusion.

Signed-off-by: Alex Lewontin <alex.lewontin@canonical.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>